Temporarily remove external contributor comment for `rapidsai` org
This PR temporarily skips the external contributor comment that gets posted for the `rapidsai` organization. This is to prevent confusion with the existing `Can one of the admins verify this patch?` comment that gets posted as part of our existing CI setup (like [this](https://github.com/rapidsai/cudf/pull/11164#issuecomment-1168999614)).

Once we're fully migrated away from the existing setup, we should revert this PR.

Also worth noting, the `okay to test` comment will still work for this `ops-bot` plugin, we just won't be posting the external contributor comment for any `rapidsai` organization repositories.