update `getPRNumbersfromSHA` implementation
The current method of using `listPullRequestsAssociatedWithCommit` to find PRs associated with a commit status is unreliable.

There are cases where it doesn't work correctly. For example, if someone other than the original PR author pushes a commit to an open PR, the current implementation will try to search that person's fork for an open PR rather than the original author's fork. This search will fail.

It's easier and more accurate to use the search API with a proper query string.